[MA-001] fixed bugs
deleted old networkManager file

updated the description of network engine

Modified ListViewModel and renamed it
